+++ packages/assignment/src/providers/NotificationProvider.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren, useState } from "react";
 import { useCallback, useMemo } from "../@lib";
 import { NotificationContext } from "../context";
-import { NotificationSystem } from "../components/NotificationSystem";
+import { NotificationSystem } from "../components";
 import { Notification } from "../types";
 
 export const NotificationProvider: React.FC<PropsWithChildren> = ({ children }) => {
